giancarlo: implementate alcune correzioni sonar. modificato il pom per poter lanciare il report di coverage di jacoco. siamo a 47.5%. mancano sotanzialmente i package gui, client e server.

+++ src/main/java/it/polimi/server/GameRoom.java
@@ -85,7 +85,7 @@ public class GameRoom extends BaseObservable implements BaseObserver{
         if(this.manager.hasOneClient()){
         	timer = new Timer();
         	TimeLimitStartGameRoom task = new TimeLimitStartGameRoom(this);
-        	timer.schedule(task, TIME_LIMIT_FOR_START*60*1000);
+        	timer.schedule(task, (long) TIME_LIMIT_FOR_START*60*1000);
         }
     }
